fix: move selectors to correct Hacksoft location
Selectors should be siblings to services/, not nested under them.
- clients/services/merge_selectors.py → clients/selectors/merge.py
- Matches existing pattern: reports/selectors.py, shelters/selectors/,
  teams/selectors.py, accounts/selectors.py
